add hideTitle param to pageDefiniton
